Rename all UBL files with a ubl_ prefix
