+++ Marlin/ubl_motion.cpp
@@ -24,7 +24,7 @@
 #if ENABLED(AUTO_BED_LEVELING_UBL)
 
   #include "Marlin.h"
-  #include "UBL.h"
+  #include "ubl.h"
   #include "planner.h"
   #include <avr/io.h>
   #include <math.h>
